+++ src/components/Avatar/AvatarContent.tsx
@@ -1,35 +1,35 @@
 import React from "react";
 import cx from "classnames";
 import { BEMClass } from "../../helpers/bem-helper";
-import { AvatarSizes, AvatarTypes } from "./AvatarConstants";
+import { AvatarSize, AvatarType } from "./AvatarConstants";
+import Icon from "../Icon/Icon";
+import { SubIcon, VibeComponentProps } from "../../types";
 import "./AvatarContent.scss";
-import VibeComponentProps from "src/types/VibeComponentProps";
-import Icon, { iconSubComponentProps } from "../Icon/Icon";
 
 const AVATAR_CONTENT_CSS_BASE_CLASS = "monday-style-avatar-content";
 const bemHelper = BEMClass(AVATAR_CONTENT_CSS_BASE_CLASS);
 
 interface AvatarContentProps extends VibeComponentProps {
   src?: string;
-  type?: AvatarTypes;
-  size?: AvatarSizes;
+  type?: AvatarType;
+  size?: AvatarSize;
   role?: string;
   ariaLabel?: string;
   /** we support two types of icons - SVG and FONT (classname) so this prop is either the name of the icon or the component */
-  icon?: string | React.FunctionComponent<iconSubComponentProps> | null;
+  icon?: SubIcon;
   textClassName?: string;
   text?: string;
 }
 
 export const AvatarContent: React.FC<AvatarContentProps> & {
-  sizes?: typeof AvatarSizes;
-  types?: typeof AvatarTypes;
-} = ({ type, src, icon, text, ariaLabel, role, size, textClassName }) => {
+  sizes?: typeof AvatarSize;
+  types?: typeof AvatarType;
+} = ({ type = AvatarType.TEXT, src, icon, text, ariaLabel, role, size = AvatarSize.LARGE, textClassName = "" }) => {
   const className = cx(bemHelper({ element: type }), bemHelper({ element: type, state: size }));
   switch (type) {
-    case AvatarTypes.IMG:
+    case AvatarType.IMG:
       return <img role={role} alt={ariaLabel} src={src} className={className} />;
-    case AvatarTypes.ICON:
+    case AvatarType.ICON:
       return (
         <Icon
           icon={icon}
@@ -40,7 +40,7 @@ export const AvatarContent: React.FC<AvatarContentProps> & {
           ariaHidden={false}
         />
       );
-    case AvatarTypes.TEXT:
+    case AvatarType.TEXT:
       return (
         <span aria-label={ariaLabel} role={role} className={cx(className, textClassName)}>
           {text}
@@ -52,17 +52,6 @@ export const AvatarContent: React.FC<AvatarContentProps> & {
 };
 
 Object.assign(AvatarContent, {
-  types: AvatarTypes,
-  sizes: AvatarSizes
+  types: AvatarType,
+  sizes: AvatarSize
 });
-
-AvatarContent.defaultProps = {
-  src: undefined,
-  icon: undefined,
-  type: AvatarTypes.TEXT,
-  role: undefined,
-  ariaLabel: undefined,
-  size: AvatarSizes.LARGE,
-  textClassName: "",
-  text: undefined
-};