+++ src/components/BreadcrumbsBar/BreadcrumbsConstants.ts
@@ -1,0 +1,4 @@
+export enum BreadcrumbsBarType {
+  NAVIGATION = "navigation",
+  INDICATION = "indication"
+}