+++ src/components/ProgressBars/LinearProgressBar/LinearProgressBarConstants.ts
@@ -1,0 +1,9 @@
+export const baseClassName = "linear-progress-bar";
+
+export enum ProgressBarStyle {
+  PRIMARY = "primary",
+  SECONDARY = "secondary",
+  POSITIVE = "positive",
+  NEGATIVE = "negative",
+  NONE = "none"
+}