+++ src/components/Search/SearchConstats.js
@@ -1,5 +1,0 @@
-export const TYPES = {
-  SQUARE: "square",
-  ROUND: "round",
-  UNDERLINE: "underline"
-};